Add files via upload
Button is use for internal pull up resistor to keep the button HIGH when not pressed. But the connection to ground when pressed making it LOW, If the button is pressed HIGH the "group1_on" when the button if not pressed LOW it will send "group1_off". Message are sent to the serial monitor every 2 seconds. And its useful for controlling the external system via serial communication.